Removed MVC project. Redesigned the models: added competition.

+++ Jbit.Common/Services/ITaskRatingCalculator.cs
@@ -9,6 +9,6 @@
         /// Calculate task raiting
         /// </summary>
         /// <returns></returns>
-        double GetRating();
+        decimal GetRating();
     }
 }
